Ueransim (#10)
* Add multiple components
---------

+++ .global/pac/DEPLOY_COMPONENT.groovy
@@ -9,33 +9,33 @@ pipeline {
 
   parameters {
     string(name: 'TN_ID', defaultValue: '', description: 'TRIAL NETWORK IDENTIFIER')
-    string(name: 'LIBRARY_COMPONENT_NAME', defaultValue: '', description: '6G LIBRARY COMPONENT')
     string(name: 'LIBRARY_BRANCH', defaultValue: 'main', description: '6G LIBRARY BRANCH')
+    string(name: 'LIBRARY_COMPONENT_NAME', defaultValue: '', description: '6G LIBRARY COMPONENT')
     string(name: 'DEPLOYMENT_SITE', defaultValue: 'uma', description: 'Site where deploy')
+    booleanParam(name: 'DEBUG', defaultValue: false, description: 'Enable DEBUG')
     base64File (name: 'FILE', description: 'YAML file that contains variables needed to deploy the component')
     //string(name: 'TNLCM_CALLBACK', defaultValue: 'https://tnlcm.uma/TN/ID/callback', description: 'URL of the TNLCM to notify result')   
   }
 
   environment {
-      LIBRARY_URL="https://github.com/6G-SANDBOX/6G-Library"
       TN_ID="${params.TN_ID}"
-      LIBRARY_COMPONENT_NAME="${params.LIBRARY_COMPONENT_NAME}"
       LIBRARY_BRANCH="${params.LIBRARY_BRANCH}"
+      LIBRARY_COMPONENT_NAME="${params.LIBRARY_COMPONENT_NAME}"
       DEPLOYMENT_SITE="${params.DEPLOYMENT_SITE}"
       FILE_PATH="${WORKSPACE}/${params.LIBRARY_COMPONENT_NAME}/private/tnlcm_vars.yaml"
+      DEBUG="${params.DEBUG}"
       OPENNEBULA_API_CREDENTIALS = credentials('OPENNEBULA_API_CREDENTIALS')
       OPENNEBULA_USERNAME = credentials('OPENNEBULA_USERNAME')
       OPENNEBULA_PASSWORD = credentials('OPENNEBULA_PASSWORD')
       OPENNEBULA_ENDPOINT = credentials('OPENNEBULA_ENDPOINT')
+      OPENNEBULA_FLOW_ENDPOINT = credentials('OPENNEBULA_FLOW_ENDPOINT')
       OPENNEBULA_INSECURE = credentials('OPENNEBULA_INSECURE')
-      // MINIO CREDENTIALS
+      ONE_XMLRPC = credentials('ONE_XMLRPC')
+      ONE_AUTH = credentials('ONE_AUTH')
       AWS_ACCESS_KEY_ID = credentials('MINIO_KEY')
       AWS_SECRET_ACCESS_KEY = credentials('MINIO_SECRET')
-      //ANSIBLE
       ANSIBLE_REMOTE_USER = credentials('ANSIBLE_REMOTE_USER')
       ANSIBLE_CONNECTION_PASSWORD_FILE = credentials('ANSIBLE_CONNECTION_PASSWORD_FILE')
-      //ANSIBLE_BECOME_PASSWORD_FILE = credentials('ANSIBLE_BECOME_PASSWORD_FILE')    
-      //GITHUB JENKINS TOKEN
       GITHUB_JENKINS = credentials('GITHUB_JENKINS')  
   }
 
@@ -48,7 +48,6 @@ pipeline {
                 script: 'echo $FILE | base64 -d',
                 returnStdout: true)
             echo "${YAML_CONTENT}"
-            // This file path is relevant because ansible will look for this file on stage 2
             writeFile(file: FILE_PATH, text: YAML_CONTENT)
           }
       }
@@ -82,16 +81,24 @@ pipeline {
           '''
       }
       cleanup{
-          /* clean up our workspace */
-          deleteDir()
-          /* clean up tmp directory */
-          dir("${env.workspace}@tmp") {
-              deleteDir()
+          script {
+              if (env.DEBUG == 'false') {
+                  echo 'Clean up Workspace'
+                  deleteDir()
+              
+                  echo 'Clean up tmp directory'
+                  dir("${env.workspace}@tmp") {
+                      deleteDir()
+                  }
+              
+                  echo 'Clean up script directory'
+                  dir("${env.workspace}@script") {
+                      deleteDir()
+                  }
+              } else {
+                  echo 'Workspace is not deleted'
+              }
           }
-          /* clean up script directory */
-          dir("${env.workspace}@script") {
-              deleteDir()
-          }
-      }      
-  }  
+      }
+  }
 }
